Force actions with prefix name to appear anyway.
If we add to the tree something called "program" and then
something else called "prog", the 2nd one is a prefix of the 1st one
and all the letters in the search tree will be taken already.

This commit switches the last letter, so that things which have
prefix names can still be accessed.

This bug arised by having .desktop files from multiple directories.
Originally the problem was not there because they are explored in
alphabetical order.