Keep separate list of Action inside the tree
A single action can be associated with more than one node
within the tree.

This complicates freeing them, because all the references but one
need to be set to NULL. However finding them can be difficult because
they are not necessarily all located on a path.

So I am keeping a separate list with all the Actions, to delete
them separately from the nodes.