Remove list of pointers to actions
Since now all the actions can be owned by a parent object, when rescans
happens, the parent object is destroyed, causing all the actions to be
automatically destroyed by Qt.

This means that there is no longer need to keep a list of pointers to the
actions, for when they need to be removed.