The global shortcut toggles the window
Instead of using the global shortcut to just show the window, it can now
be used to also hide the window.

This is convenient because, if some other window pops up and gets the focus
away from trabucco, a click was usually necessary to go back to its window.

Now if this happens, it is possible to press the global shortcut twice to
get the window focused and in front.