Rewrite self-reloading code
The previous one was not working because too
many resources remained open after closing the QApplication.

This closes the entire process so all the resources are freed.

An additional complication comes from the fact that trabucco only allows
one instance per session, so a --wait parameter is added, and if that is
present, some more time is given to aquire the lock.